requirements.txt
#Core lib
hydra-core
python-dateutil
schedule
PyJWT
argh

# Connectors
sqlalchemy
alembic
pymysql
psycopg2-binary
pysolr
pymongo
python-memcached
cryptography
boto3

# web
django
flask


# Test
python-dotenv